chg: Some refactoring of database queries
No changes in logic, but it does gert rid of some branching.
