+++ actix/src/database.rs
@@ -1,16 +1,15 @@
 // SPDX-FileCopyrightText: 2023 Sayantan Santra <sayantan.santra689@gmail.com>
 // SPDX-License-Identifier: MIT
 
-use log::{debug, error, info};
-use rusqlite::{Connection, fallible_iterator::FallibleIterator, named_params, types::Value};
+use log::error;
+use rusqlite::{Connection, fallible_iterator::FallibleIterator, named_params};
 use serde::Serialize;
-use std::{collections::HashSet, fs, rc::Rc};
+use std::rc::Rc;
+
+mod strings;
+pub mod utils;
 
 use crate::services::ChhotoError::{self, ClientError, ServerError};
-
-// Some constants
-const APPLICATION_ID: i32 = i32::from_be_bytes(*b"chht"); // MUST NEVER BE CHANGED
-const USER_VERSION: u32 = 4; // Should be incremented on change of schema
 
 // Struct for encoding a DB row
 #[derive(Serialize)]
@@ -22,22 +21,11 @@ pub struct DBRow {
     pub notes: String,
 }
 
-// Struct for creating a get query
-struct QueryHelper {
-    prefix: String,
-    joins: String,
-    conditions: String,
-    suffix: String,
-}
-
 // Find a single URL for /api/expand
 pub fn find_url(shortlink: &str, db: &Connection) -> Result<DBRow, ChhotoError> {
     // Long link, hits, expiry time
     let now = chrono::Utc::now().timestamp();
-    let query = "SELECT long_url, hits, expiry_time, notes FROM urls
-                 WHERE short_url = :short
-                 AND (expiry_time IS NULL OR expiry_time > :now)";
-    let Ok(mut statement) = db.prepare_cached(query) else {
+    let Ok(mut statement) = db.prepare_cached(strings::FIND_URL) else {
         error!("Error preparing SQL statement for find_url.");
         return Err(ServerError);
     };
@@ -65,68 +53,63 @@ pub fn getall(
     filter: Option<String>,
 ) -> Rc<[DBRow]> {
     let now = chrono::Utc::now().timestamp();
-    let size = page_size.unwrap_or(10);
 
-    let mut params: Vec<(&str, Value)> = vec![(":now", now.into())];
+    let has_cursor = page_after.is_some();
+    let has_filter = filter.is_some();
 
-    let mut query_helper = if let Some(pos) = page_after.as_ref() {
-        params.push((":pos", (*pos).to_owned().into()));
-        params.push((":size", size.into()));
-        QueryHelper {
-            prefix: "( SELECT t.id, t.short_url, t.long_url, t.hits, t.expiry_time, t.notes FROM urls AS t".to_string(),
-            joins: "JOIN urls AS u ON u.short_url = :pos".to_string(),
-            conditions: "WHERE t.id < u.id AND ( t.expiry_time IS NULL OR t.expiry_time > :now".to_string(),
-            suffix: ") ORDER BY t.id DESC LIMIT :size ) as t".to_string(),
-        }
-    } else if let Some(num) = page_no.as_ref() {
-        let page = (num - 1) * size;
-        params.push((":page", page.into()));
-        params.push((":size", size.into()));
-        QueryHelper {
-            prefix: "( SELECT t.id, t.short_url, t.long_url, t.hits, t.expiry_time, t.notes FROM urls AS t".to_string(),
-            joins: String::new(),
-            conditions: "WHERE ( t.expiry_time IS NULL OR t.expiry_time > :now )".to_string(),
-            suffix: "ORDER BY t.id DESC LIMIT :size OFFSET :page ) as t".to_string(),
-        }
-    } else if page_size.is_some() {
-        params.push((":size", size.into()));
-        QueryHelper {
-            prefix: "( SELECT t.id, t.short_url, t.long_url, t.hits, t.expiry_time, t.notes FROM urls AS t".to_string(),
-            joins: String::new(),
-            conditions: "WHERE ( t.expiry_time IS NULL OR t.expiry_time > :now )".to_string(),
-            suffix: "ORDER BY t.id DESC LIMIT :size ) as t".to_string(),
-        }
+    let size = if has_cursor || page_no.is_some() {
+        page_size.unwrap_or(10)
     } else {
-        QueryHelper {
-            prefix: "urls AS t".to_string(),
-            joins: String::new(),
-            conditions: "WHERE ( t.expiry_time IS NULL OR t.expiry_time > :now )".to_string(),
-            suffix: String::new(),
-        }
+        i64::MAX
+    };
+    let offset = page_no.map(|n| (n - 1) * size).unwrap_or(0);
+
+    let (query, params) = match (has_cursor, has_filter) {
+        (false, false) => (
+            strings::GETALL_QUERIES[0],
+            named_params! {
+                ":now": now,
+                ":size": size,
+                ":offset": offset,
+            },
+        ),
+        (true, false) => (
+            strings::GETALL_QUERIES[1],
+            named_params! {
+                ":now": now,
+                ":size": size,
+                ":pos": page_after,
+            },
+        ),
+        (false, true) => (
+            strings::GETALL_QUERIES[2],
+            named_params! {
+                ":now": now,
+                ":size": size,
+                ":offset": offset,
+                ":filter": filter,
+            },
+        ),
+        (true, true) => (
+            strings::GETALL_QUERIES[3],
+            named_params! {
+                ":now": now,
+                ":size": size,
+                ":pos": page_after,
+                ":filter": filter,
+            },
+        ),
     };
 
-    if let Some(fil) = filter {
-        query_helper
-            .joins
-            .push_str(" JOIN urls_fts AS f ON t.id = f.rowid");
-        query_helper
-            .conditions
-            .push_str(" AND urls_fts MATCH :filter");
-        params.push((":filter", fil.into()));
-    }
-
-    let query = format!(
-        "SELECT t.short_url, t.long_url, t.hits, t.expiry_time, t.notes FROM {} {} {} {} ORDER BY t.id ASC",
-        query_helper.prefix, query_helper.joins, query_helper.conditions, query_helper.suffix,
-    );
-    let Ok(mut statement) = db.prepare_cached(&query) else {
+    let Ok(mut statement) = db.prepare_cached(query) else {
         error!("Error preparing SQL statement for getall.");
         return [].into();
     };
-    let raw_data = statement.query(params.as_slice());
+
+    let raw_data = statement.query(params);
 
     let Ok(data) = raw_data else {
-        error!("Error running SQL statement for getall: {query}");
+        error!("Error running SQL statement for getall.");
         return [].into();
     };
 
@@ -152,12 +135,7 @@ pub fn getall(
 // Add a hit when site is visited during link resolution
 pub fn find_and_add_hit(shortlink: &str, db: &Connection) -> Result<String, ()> {
     let now = chrono::Utc::now().timestamp();
-    let Ok(mut statement) = db.prepare_cached(
-        "UPDATE urls 
-             SET hits = hits + 1 
-             WHERE short_url = :short AND (expiry_time IS NULL OR expiry_time > :now)
-             RETURNING long_url",
-    ) else {
+    let Ok(mut statement) = db.prepare_cached(strings::FIND_AND_ADD_HIT) else {
         error!("Error preparing SQL statement for add_hit.");
         return Err(());
     };
@@ -179,26 +157,22 @@ pub fn add_link(
     let now = chrono::Utc::now().timestamp();
     let expiry_time = expiry_delay.map(|delay| now + delay);
 
-    let Ok(mut statement) = db.prepare_cached(
-        "INSERT INTO urls
-             (long_url, short_url, hits, expiry_time, notes)
-             VALUES (:long, :short, 0, :expiry, :notes)
-             ON CONFLICT(short_url) DO UPDATE 
-             SET long_url = :long, hits = 0, expiry_time = :expiry, notes = :notes
-             WHERE short_url = :short AND expiry_time <= :now AND expiry_time IS NOT NULL",
-    ) else {
+    let Ok(mut statement) = db.prepare_cached(strings::ADD_LINK) else {
         error!("Error preparing SQL statement for add_link.");
         return Err(ServerError);
     };
-    match statement.execute(
-        named_params! {":long": longlink, ":short": shortlink, ":expiry": expiry_time, ":now": now, ":notes" : notes.filter(|s| !s.is_empty())},
-    ) {
+    match statement.execute(named_params! {":long": longlink, ":short": shortlink,
+    ":expiry": expiry_time, ":now": now, ":notes" : notes.filter(|s| !s.is_empty())})
+    {
         Ok(1) => Ok(expiry_time.unwrap_or_default()),
         Ok(_) => Err(ClientError {
             reason: "Short URL is already in use!".to_string(),
         }),
         Err(e) => {
-            error!("There was some error while adding the link ({shortlink}, {longlink}, {:?}): {e}", expiry_delay);
+            error!(
+                "There was some error while adding the link ({shortlink}, {longlink}, {:?}): {e}",
+                expiry_delay
+            );
             Err(ServerError)
         }
     }
@@ -214,41 +188,19 @@ pub fn edit_link(
     db: &Connection,
 ) -> Result<usize, ()> {
     let now = chrono::Utc::now().timestamp();
-    let mut params: Vec<(&str, Value)> = vec![
-        (":long", longlink.to_owned().into()),
-        (":short", shortlink.to_owned().into()),
-        (":now", now.into()),
-    ];
-
-    let mut updates = "long_url = :long".to_string();
-    if reset_hits {
-        updates.push_str(", hits = 0")
-    };
-    if let Some(note) = notes.as_ref()
-        && !note.is_empty()
-    {
-        params.push((":notes", (*note).to_owned().into()));
-        updates.push_str(", notes = :notes");
-    };
-    if let Some(&expiry) = expiry_time.as_ref()
-        && expiry > 0
-    {
-        params.push((":expiry", expiry.into()));
-        updates.push_str(", expiry_time = :expiry")
-    };
-
-    let query = format!(
-        "UPDATE urls
-         SET {updates}
-         WHERE short_url = :short AND (expiry_time IS NULL OR expiry_time > :now)"
-    );
-    let Ok(mut statement) = db.prepare_cached(&query) else {
+    let Ok(mut statement) = db.prepare_cached(strings::EDIT_LINK) else {
         error!("Error preparing SQL statement for edit_link.");
         return Err(());
     };
-
     statement
-        .execute(params.as_slice())
+        .execute(named_params! {
+            ":long": longlink,
+            ":short": shortlink,
+            ":now": now,
+            ":hits": reset_hits.then_some(0),
+            ":notes": notes,
+            ":expiry": expiry_time,
+        })
         .inspect_err(|err| {
             error!(
                 "Got an error while editing link ({shortlink}, {longlink}, {reset_hits}): {err}"
@@ -257,54 +209,9 @@ pub fn edit_link(
         .map_err(|_| ())
 }
 
-// Clean expired links
-pub fn cleanup(db: &Connection, use_wal_mode: bool) {
-    let now = chrono::Utc::now().timestamp();
-    debug!("Starting database cleanup.");
-
-    db.prepare_cached("DELETE FROM urls WHERE :now >= expiry_time AND expiry_time IS NOT NULL")
-        .expect("Error preparing SQL statement for cleanup.")
-        .execute(named_params! {":now" : now})
-        .inspect(|&u| match u {
-            0 => (),
-            1 => info!("1 expired link was deleted."),
-            _ => info!("{u} expired links were deleted."),
-        })
-        .expect("Error cleaning expired links.");
-
-    if use_wal_mode {
-        let mut statement = db
-            .prepare_cached("PRAGMA wal_checkpoint(PASSIVE)")
-            .expect("Error preparing SQL statement for pragma: wal_checkpoint.");
-        statement
-            .query_one([], |row| row.get::<usize, isize>(1))
-            .ok()
-            .filter(|&v| v != -1)
-            .expect("Unable to create WAL checkpoint.");
-    }
-    let freelist_count: i64 = db
-        .prepare_cached("PRAGMA freelist_count")
-        .expect("Error preparing SQL statement for pragma: freelist_count")
-        .query_row([], |r| r.get(0))
-        .expect("failed to get freelist_count");
-
-    // Roughly 20 MB with 4 KiB pages
-    if freelist_count > 5000 {
-        db.prepare_cached("VACUUM")
-            .expect("Error preparing SQL statement for vacuum.")
-            .execute([])
-            .expect("failed to vacuum database");
-    }
-    db.prepare_cached("PRAGMA optimize")
-        .expect("Error preparing SQL statement for pragma: optimize.")
-        .execute([])
-        .expect("Unable to optimize database.");
-    debug!("Optimized database.")
-}
-
 // Delete an existing link
 pub fn delete_link(shortlink: &str, db: &Connection) -> Result<(), ChhotoError> {
-    let Ok(mut statement) = db.prepare_cached("DELETE FROM urls WHERE short_url = :short") else {
+    let Ok(mut statement) = db.prepare_cached(strings::DELETE_LINK) else {
         error!("Error preparing SQL statement for delete_link.");
         return Err(ServerError);
     };
@@ -314,252 +221,4 @@ pub fn delete_link(shortlink: &str, db: &Connection) -> Result<(), ChhotoError> 
             reason: "The shortlink was not found, and could not be deleted.".to_string(),
         }),
     }
-}
-
-// Initialize the database
-pub fn initialize_db(path: &str, use_wal_mode: bool, ensure_acid: bool) {
-    let mut db = Connection::open(path).expect("Unable to open database!");
-
-    info!("Creating a backup of the existing database.");
-    let bak1 = format!("{path}.bak1");
-    let bak2 = format!("{path}.bak2");
-    if fs::exists(&bak1).unwrap_or(false) {
-        fs::rename(&bak1, &bak2).expect("Error while renaming old backup.");
-    }
-    db.backup("main", &bak1, None)
-        .expect("Error while creating backup.");
-
-    info!("Initializing database.");
-    let (mut tables, mut indices) = db
-        .prepare(
-            "SELECT type, name
-             FROM sqlite_master
-             WHERE type IN ('table', 'index') AND name NOT LIKE 'sqlite_%'",
-        )
-        .expect("Error preparing statement for database objects query.")
-        .query_map([], |row| {
-            Ok((row.get::<_, String>("type")?, row.get::<_, String>("name")?))
-        })
-        .expect("Error executing database objects query.")
-        .filter_map(Result::ok)
-        .fold(
-            (HashSet::new(), HashSet::new()),
-            |(mut tables, mut indices), (obj_type, name)| {
-                match obj_type.as_str() {
-                    "table" => {
-                        tables.insert(name);
-                    }
-                    "index" => {
-                        indices.insert(name);
-                    }
-                    _ => {}
-                }
-
-                (tables, indices)
-            },
-        );
-
-    let urls_table_exists = tables.contains("urls");
-    let urls_fts_table_exists = tables.contains("urls_fts");
-
-    let current_user_version: u32 = if urls_table_exists {
-        db.query_row_and_then("SELECT user_version FROM pragma_user_version", (), |row| {
-            row.get(0)
-        })
-        .unwrap_or_default()
-    } else {
-        USER_VERSION
-    };
-
-    let current_application_id: i32 = db
-        .query_row_and_then(
-            "SELECT application_id FROM pragma_application_id",
-            [],
-            |row| row.get(0),
-        )
-        .unwrap_or_default();
-    if current_application_id != 0
-        || (urls_table_exists && urls_fts_table_exists && current_user_version > 1)
-    {
-        assert_eq!(
-            current_application_id, APPLICATION_ID,
-            "Incorrect application_id: The database file seems to belong to some other application."
-        )
-    } else {
-        db.pragma_update(None, "application_id", APPLICATION_ID)
-            .expect("Unable to set pragma: application_id.");
-    }
-    let urls_table_schema = "CREATE TABLE urls (
-                                 id INTEGER PRIMARY KEY,
-                                 short_url TEXT NOT NULL,
-                                 long_url TEXT NOT NULL,
-                                 hits INTEGER NOT NULL,
-                                 expiry_time INTEGER,
-                                 notes TEXT
-                             )";
-    // Create table if it doesn't exist
-    if !urls_table_exists {
-        info!("Creating an empty urls table.");
-        db.execute(
-            urls_table_schema,
-            // expiry_time is added later during migration 1
-            [],
-        )
-        .expect("Unable to initialize empty database.");
-    }
-
-    // Migration 1: Add expiry_time, introduced in 6.0.0
-    if current_user_version < 1 {
-        info!("Applying migration 1: Add expiry_time column to urls.");
-        let tx = db
-            .transaction()
-            .expect("Unable to create transaction for migration 1.");
-        tx.execute(
-            "ALTER TABLE urls ADD COLUMN expiry_time INTEGER NOT NULL DEFAULT 0",
-            [],
-        )
-        .expect("Unable to apply migration 1.");
-        tx.pragma_update(None, "user_version", 1)
-            .expect("Unable to set pragma: user_version.");
-        tx.commit()
-            .expect("Unable to commit transaction for migration 1.");
-    }
-    // Migration 2: Add notes, introduced in 7.0.0
-    if current_user_version < 3 {
-        info!("Applying migration 2: Add notes column to urls.");
-        let tx = db
-            .transaction()
-            .expect("Unable to create transaction for migration 2.");
-        tx.execute("ALTER TABLE urls ADD COLUMN notes TEXT", ())
-            .expect("Unable to apply migration 2.");
-        tx.pragma_update(None, "user_version", 2)
-            .expect("Unable to set pragma: user_version.");
-        tx.commit()
-            .expect("Unable to commit transaction for migration 2.");
-    }
-    // Migration 3: Remove AUTOINCREMENT from the id row
-    if current_user_version < 4 {
-        info!("Applying migration 3: Remove AUTOINCREMENT from id row.");
-        let tx = db
-            .transaction()
-            .expect("Unable to create transaction for migration 2.");
-        tx.execute("ALTER TABLE urls RENAME TO urls_old", ())
-            .expect("Unable to temporarily rename urls to urls_old.");
-        tx.execute(urls_table_schema, ())
-            .expect("Unable to create new urls table.");
-        tx.execute(
-            "INSERT INTO urls (long_url, short_url, hits, expiry_time, notes)
-             SELECT long_url, short_url, hits, NULLIF(expiry_time,0), NULLIF(notes,'')
-             FROM urls_old       
-             ORDER BY id",
-            (),
-        )
-        .expect("Unable to clone data to the new table.");
-        tx.execute("DROP TABLE urls_old", ())
-            .expect("Unable to delete old urls table.");
-        if urls_fts_table_exists {
-            tx.execute("DROP TABLE urls_fts", ())
-                .expect("Unable to delete old urls_fts table.");
-        }
-        (tables, indices) = (HashSet::from(["urls".to_string()]), HashSet::new());
-        tx.pragma_update(None, "user_version", 3)
-            .expect("Unable to set pragma: user_version.");
-        tx.commit()
-            .expect("Unable to commit transaction for migration 3.");
-        db.execute("VACUUM", ())
-            .expect("failed to vacuum database after migration 3.");
-    }
-
-    // Create index on short_url for faster lookups
-    if !indices.contains("idx_short_url") {
-        info!("Creating index idx_short_url on urls(short_url).");
-        db.execute("CREATE UNIQUE INDEX idx_short_url ON urls (short_url)", ())
-            .expect("Unable to create index on short_url.");
-    }
-
-    // Create index on expiry_time for faster lookups
-    if !indices.contains("idx_expiry_time") {
-        info!("Creating index idx_expiry_time on urls(expiry_time).");
-        db.execute("CREATE INDEX idx_expiry_time ON urls (expiry_time)", ())
-            .expect("Unable to create index on expiry_time.");
-    }
-
-    // Create FTS5 table if it doesn't exist, and also create triggers
-    if !tables.contains("urls_fts") {
-        info!("Creating FTS table urls_fts, and adding triggers.");
-        let tx = db
-            .transaction()
-            .expect("Unable to create transaction for FTS table creation.");
-        tx.execute(
-            "CREATE VIRTUAL TABLE urls_fts USING fts5(
-             long_url, short_url, notes,
-             content='urls',
-             content_rowid='id',
-             tokenize='trigram remove_diacritics 2'
-         )",
-            [],
-        )
-        .expect("Unable to create FTS table.");
-
-        tx.execute("INSERT INTO urls_fts(urls_fts) VALUES ('rebuild')", ())
-            .expect("Unable to populate FTS table.");
-        let fts_triggers = [
-            "CREATE TRIGGER urls_insert
-                 AFTER INSERT ON urls BEGIN
-                 INSERT INTO urls_fts(rowid, long_url, short_url, notes)
-                 VALUES (new.id, new.long_url, new.short_url, new.notes);
-             END",
-            "CREATE TRIGGER urls_delete
-                 AFTER DELETE ON urls BEGIN
-                 INSERT INTO urls_fts(urls_fts, rowid, long_url, short_url, notes)
-                 VALUES('delete', old.id, old.long_url, old.short_url, old.notes);
-             END",
-            "CREATE TRIGGER urls_update
-             AFTER UPDATE ON urls BEGIN
-                 INSERT INTO urls_fts(urls_fts, rowid, long_url, short_url, notes)
-                 VALUES('delete', old.id, old.long_url, old.short_url, old.notes);
-                 INSERT INTO urls_fts(rowid, long_url, short_url, notes)
-                 VALUES (new.id, new.long_url, new.short_url, new.notes);
-             END",
-        ];
-        for trigger in fts_triggers {
-            tx.execute(trigger, ())
-                .expect("Unable to create FTS trigger(s).");
-        }
-
-        tx.commit().expect("Unable to create FTS table.");
-    }
-
-    // Set WAL mode if specified
-    let (journal_mode, synchronous) = match (use_wal_mode, ensure_acid) {
-        (true, false) => ("WAL", "NORMAL"),
-        (true, true) => ("WAL", "FULL"),
-        (false, false) => ("DELETE", "FULL"),
-        (false, true) => ("DELETE", "EXTRA"),
-    };
-    db.pragma_update(None, "journal_mode", journal_mode)
-        .expect("Unable to set pragma: journal_mode.");
-    db.pragma_update(None, "synchronous", synchronous)
-        .expect("Unable to set pragma: synchronous.");
-    db.pragma_update(None, "temp_store", "memory")
-        .expect("Unable to set pragma: temp_store.");
-    let tx = db
-        .transaction()
-        .expect("Unable to create transaction for pragma updates.");
-    // Set some further optimizations and run vacuum if necessary
-    tx.pragma_update(None, "journal_size_limit", "8388608")
-        .expect("Unable to set pragma: journal_size_limit.");
-    tx.pragma_update(None, "mmap_size", "16777216")
-        .expect("Unable to set pragma: mmap_size.");
-    // The schema should be up-to-date by this point
-    tx.pragma_update(None, "user_version", USER_VERSION)
-        .expect("Unable to set pragma: user_version.");
-    tx.commit().expect("Unable to set correct pragma values.");
-
-    info!("Database initialization was successful.");
-}
-
-// Open and return a rusqlite connection
-pub fn open_db(path: &str) -> Connection {
-    Connection::open(path).expect("Unable to open database.")
 }
